test: add custom eslint rule tests
