+++ tools/eslint/package.json
@@ -1,5 +1,8 @@
 {
   "name": "eslint-plugin-midas",
   "version": "1.0.0",
-  "main": "index.js"
+  "main": "index.js",
+  "scripts": {
+    "test": "node handle-deprecated-comments.test.js"
+  }
 }
